[feat] #4 회원 삭제 및 회원 닉네임 수정 구현

+++ src/main/java/com/example/umc3_teamproject/dto/SignupRes.java
@@ -9,6 +9,6 @@ import lombok.Setter;
 @AllArgsConstructor
 public class SignupRes {
 
-    private Long memberId;
+    private int memberId;
     private String jwt;
 }
